requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.5
django-cleanup==5.1.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
openpyxl==3.0.7
Pillow==8.1.0
psycopg2-binary==2.8.6
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
